Revamps project including build
This moves the sender-specific logic to configuration, as we should be
able to re-use config and tests in zipkin-reporter better this way. This
leaves a couple ITs for historical sake.

While all of the logic has been rewritten, the spring configuration is
the same. Notably, this is flattened into a single module and single
Docker image as it is much easier to deal with this way, and pretty much
identical to zipkin-storage-kafka now.

Finally, this moves to GitHub actions publishing (no more bintray, too)

Enjoy!